** MODIFY VALUATION SCHEME **
so that children are now counted as many times as they appear.
The side effect of this commit is that,
now duplicate 'Vert' objects are counted multiple times.
This was actually intended all along.

This causes a massive increase in speed,
and is what we have been working towards.
Going forwards, we will be able to delete a lot of redundant code.